Cleaned up our use of some mappings.
