+++ src/analyser.cpp
@@ -34,6 +34,7 @@ limitations under the License.
 #include "analysermodel_p.h"
 #include "analyservariable_p.h"
 #include "anycellmlelement_p.h"
+#include "commonutils.h"
 #include "generator_p.h"
 #include "issue_p.h"
 #include "logger_p.h"
