+++ src/generator.cpp
@@ -27,6 +27,7 @@ limitations under the License.
 #include "libcellml/units.h"
 #include "libcellml/version.h"
 
+#include "commonutils.h"
 #include "generator_p.h"
 #include "generatorprofilesha1values.h"
 #include "generatorprofiletools.h"
