+++ src/printer.cpp
@@ -33,6 +33,7 @@ limitations under the License.
 #include "libcellml/variable.h"
 
 #include "anycellmlelement_p.h"
+#include "commonutils.h"
 #include "internaltypes.h"
 #include "issue_p.h"
 #include "logger_p.h"
