+++ src/utilities.h
@@ -25,7 +25,6 @@ limitations under the License.
 #include "libcellml/types.h"
 #include "libcellml/variable.h"
 
-#include "commonutils.h"
 #include "internaltypes.h"
 
 namespace libcellml {
@@ -122,31 +121,6 @@ static const std::map<std::string, double> standardMultiplierList = {
     {"volt", 0.0},
     {"watt", 0.0},
     {"weber", 0.0}};
-
-/**
- * Map connecting prefix strings to their exponent (eg: "kilo" -> 10^3).
- */
-static const std::map<std::string, int> standardPrefixList = {
-    {"yotta", 24},
-    {"zetta", 21},
-    {"exa", 18},
-    {"peta", 15},
-    {"tera", 12},
-    {"giga", 9},
-    {"mega", 6},
-    {"kilo", 3},
-    {"hecto", 2},
-    {"deca", 1},
-    {"deci", -1},
-    {"centi", -2},
-    {"milli", -3},
-    {"micro", -6},
-    {"nano", -9},
-    {"pico", -12},
-    {"femto", -15},
-    {"atto", -18},
-    {"zepto", -21},
-    {"yocto", -24}};
 
 /**
  * List of MathML elements supported by CellML.
